+++ src/app/modules/variable_pay/entities/criteria-target.entity.ts
@@ -13,7 +13,7 @@ export class CriteriaTarget extends BaseModel {
   @ManyToOne(() => VpCriteria, { onDelete: 'CASCADE' })
   vpCriteria: VpCriteria;
 
-  @Column({ type: 'decimal', precision: 10, scale: 2, default: 1 })
+  @Column({ type: 'decimal', precision: 16, scale: 2, default: 1 })
   target: number;
 
   @Column({ type: 'varchar', nullable: true })
